+++ src/geneflow/workflow.py
@@ -355,9 +355,25 @@ class Workflow:
                 self._workflow['parameters'][parameter_key]['value']\
                     = self._job['parameters'][parameter_key]
 
-        # update final output
-        if self._job['final_output']:
-            self._workflow['final_output'] = self._job['final_output']
+        # update publish list
+        if self._job['publish']:
+            # over-ride the workflow publish list with the job publish list
+            self._workflow['publish'] = self._job['publish']
+
+        # update the publish list based on publish flag of each step
+        for step_name, step in self._workflow['steps'].items():
+            if step['publish']:
+                if step_name not in self._workflow['publish']:
+                    self._workflow['publish'].append(step_name)
+
+        #for step_name in self._workflow['publish']:
+        #    Log.some().info('publish %s', step_name)
+        #    if step_name not in self._workflow['steps']:
+        #        Log.a().warning('invalid step name in publish list: %s', step_name)
+        #    else:
+        #        self._workflow['steps'][step_name]['publish'] = True
+        import pprint
+        pprint.pprint(self._workflow['publish'])
 
         # insert step execution parameters
         for step_name, step in self._workflow['steps'].items():